Restructure for KSP support
